requirements.txt
crewai>=0.28.8
langchain-openai>=0.1.7
streamlit>=1.34.0
python-dotenv>=1.0.1
requests>=2.31.0
